sharding/requirements.txt
blinker==1.9.0
click==8.2.2
colorama==0.4.6
Flask==3.1.1
hash-ring==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
